Use fewer registers when pushing a frame.
If the `self` passed in to the frame is a temporary register, we can
reuse it for other things.  I feel bad for doing this because it's
clobbering the register of the caller, but pushing a frame already
automatically released the register which is kind of also clobbering.

Maybe we should add a "freeze" to temporary register objects to prevent
clobbering?