Use Fiddle::Handle::DEFAULT as much as we can
I noticed that if you try building Ruby with debugging enabled, the
symbol addresses we extract seem to be wrong.  I think we should rely on
`dlsym` as much as we can because we know it's correct.  I'll try to
figure out what's wrong with symbol resolution, but this should fix many
problems for now.